Refactor store handler selection and registration.
The consumer can now override one of two methods. The first, `redrawOn`
specifies the stores to be listened to. The default action is to redraw the
component. The second method, `getStoreHandlers` returns a map of stores to
their handlers. The result of this method adds to / replaces the default
handlers created from `redrawOn`.

The consumer *may* choose to override both of these methods, such as when
several stores within a composite store will use the default handler, but one
needs a special handler (specified alone in `getStoreHandlers`).